pyproject.toml
[project]
name = "menu-extraction-ai"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "jinja2>=3.1.6",
    "openai>=2.15.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pymysql>=1.1.2",
    "sqlalchemy>=2.0.45",
]
